organize menu dto
