+++ src/menu/dto/create-menu.dto.ts
@@ -1,7 +1,11 @@
-import { MenuDataDto } from './menu-data.dto';
-
 export class CreateMenuDto {
   merchantId: string;
   restaurantId: string;
   data: MenuDataDto; // TODO
+}
+
+export class MenuDataDto {
+  name: string;
+  isActive: boolean;
+  index: number;
 }
