+++ src/menu/dto/index.ts
@@ -1,9 +1,9 @@
 export { GetMenuInformationDto } from './get-menu-information.dto';
 export { GetMenuItemToppingDto } from './get-menu-item-topping.dto';
-export { MenuDto } from './menu.dto';
+export { MenuDto } from './menu/menu.dto';
 export { CreateMenuDto } from './create-menu.dto';
 export { FetchMenuOfRestaurantDto } from './fetch-menu-of-restaurant.dto';
 export { UpdateMenuDto, UpdatedMenuDataDto } from './update-menu.dto';
 export { FetchMenuGroupsAndItemsDto } from './fetch-menu-groups-and-items.dto';
-export { MenuWithMenuGroupsAndItemsDto } from './menu-with-menu-groups-and-items.dto';
-export { MenuGroupWithMenuItemsDto } from './menu-group-with-menu-items.dto';
+export { MenuWithMenuGroupsAndItemsDto } from './menu-with-menu-groups-and-items/menu-with-menu-groups-and-items.dto';
+export { MenuForOrderDto } from './menu-for-order/menu-for-order.dto';
