--- src/menu/dto/menu-data.dto.ts
@@ -1,5 +1,0 @@
-export class MenuDataDto {
-  name: string;
-  isActive: boolean;
-  index: number;
-}
