+++ src/menu/dto/menu-with-menu-groups-and-items/menu-group-with-menu-items.dto.ts
@@ -1,4 +1,4 @@
-import { MenuGroup } from './../entities/menu-group.entity';
+import { MenuGroup } from '../../entities/';
 export class MenuGroupWithMenuItemsDto {
   id: string;
   name: string;
