+++ src/menu/interfaces/fetch-menu-groups-and-items-response.interface.ts
@@ -3,5 +3,7 @@ import { MenuWithMenuGroupsAndItemsDto } from '../dto';
 export interface IFetchMenuGroupsAndItemsResponse {
   status: number;
   message: string;
-  data: MenuWithMenuGroupsAndItemsDto;
+  data: {
+    menu: MenuWithMenuGroupsAndItemsDto;
+  };
 }
