+++ src/menu/interfaces/fetch-menu-of-restaurant.interface.ts
@@ -1,4 +1,4 @@
-import { MenuDto } from './../dto/menu.dto';
+import { MenuDto } from '../dto/menu/menu.dto';
 export interface IFetchMenuOfRestaurantResponse {
   status: number;
   message: string;
